+++ src/teams/repositories/xpub-team.test.ts
@@ -1,7 +1,7 @@
 import * as Knex from 'knex';
 import { MockKnex } from '../../test-mocks/knex-mock';
 import { v4 } from 'uuid';
-import { TeamId } from '../../submission/types';
+import { TeamId } from '../types';
 import XpubTeamRepository from './xpub-team';
 
 const entryId1 = TeamId.fromUuid(v4());
